move lang toggle to its own file
